requirements.txt
streamlit
requests
python-dotenv
uvicorn
fastapi
pydantic
fpdf
python-docx
langchain
langchain-google-genai
google-generativeai 